pyproject.toml
[tool.poetry]
name = "nea-maze-game"
version = "0.1.0"
description = "Non-examination-assessment_maze_game"
authors = ["JemimahOlufemi"]
readme = "README.md"
packages = [{include = "nea_maze_game"}]

[tool.poetry.dependencies]
python = "^3.10"
Panda3D = "1.10.13"
Pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
pytest = "^7.2.1"
black = "^23.1.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
